+++ Changelog.md
@@ -29,6 +29,8 @@ Deprecations
   a `after(:all)` hook. (Myron Marston, Jon Rowe)
 * Deprecate built-in `its` usage in favor of `rspec-it` gem due to planned
   removal in RSpec 3 (Peter Alfvin)
+* Deprecate `RSpec::Core::PendingExampleFixedError` in favor of
+  `RSpec::Core::Pending::PendingExampleFixedError` (Myron Marston).
 
 ### 2.14.7 / 2013-10-29
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.14.6...v2.14.7)